Updates to support v2.1.0 and up of Libseccomp
Adds a function to the public API to retrieve the library version the bindings
were built against.

Many comments and error messages throughout receive slight updates to fix
warnings identified by Golint. In general, all error messages begin with
lowercase letters and do not end in punctuation, and all comments on exported
functions and constants begin with the name of the function/variable.

Signed-off-by: Matthew Heon <mheon@redhat.com>